// src/interface/global.interface.ts
export interface Publication {
  published?: boolean
  lastPublished?: string
  presentedBy?: string
  summary?: string
  contact?: PublicationContact
  options?: PublicationOptions
}

export interface PublicationContact {
  name: string
  email: string
  phoneNumber?: string
  phone?: number | string
  formattedPhone?: number | string
}

export interface PublicationOptions {
  contactEmail: boolean
  contactName: boolean
  contactPhone: boolean
  modelName: boolean
  modelSummary: boolean
  presentedBy: boolean
  address: boolean
  externalUrl: boolean
}